requirements.txt
discord.py
python-dotenv
fuzzywuzzy
python-Levenshtein
requests
beautifulsoup4
aiocron
psycopg2
pandas
sqlalchemy
tabulate
disputils